Move file system functions to own file
