+++ library/start_menu_helper.py
@@ -1,12 +1,16 @@
 """Reorganize the start menu folder."""
 import logging
-import pathlib
 import re
 import time
-from typing import List
 
 from library import configuration, constants
 from library.helpers import windows_shortcuts
+from library.helpers.file_system import (
+    get_nested_directories,
+    get_nested_files,
+    get_nested_links,
+    resolve_files
+)
 from library.helpers.stopable_thread import StoppableThread
 
 
@@ -217,50 +221,3 @@ def delete_links_to_folders() -> None:
             if windows_shortcuts.read_shortcut(link).is_dir():
                 link.unlink()
                 logging.info(f"Deleted link to folder: {str(link)}")
-
-
-def get_nested_directories(directory: pathlib.WindowsPath) -> List[pathlib.WindowsPath]:
-    """Return all directories inside directory and its child directories."""
-    nested_directories = []
-    for item in directory.iterdir():
-        if item.is_dir():
-            nested_directories.append(item)
-    for nested_directory in nested_directories:
-        for item in nested_directory.iterdir():
-            if item.is_dir():
-                nested_directories.append(item)
-    return nested_directories
-
-
-def get_nested_files(directory: pathlib.WindowsPath) -> List[pathlib.WindowsPath]:
-    """Return all files inside directory and its child directories."""
-    files = [item for item in directory.iterdir() if item.is_file()]
-
-    for current_directory in get_nested_directories(directory):
-        for item in current_directory.iterdir():
-            if item.is_file():
-                files.append(item)
-    return files
-
-
-def get_nested_links(directory: pathlib.WindowsPath) -> List[pathlib.WindowsPath]:
-    """Return all links inside directory and its child directories."""
-    links = [item for item in directory.iterdir() if
-             item.is_symlink() or windows_shortcuts.is_shortcut(item)]
-
-    for current_directory in get_nested_directories(directory):
-        for item in current_directory.iterdir():
-            if item.is_symlink() or windows_shortcuts.is_shortcut(item):
-                links.append(item)
-    return links
-
-
-def resolve_files(files: List[pathlib.WindowsPath]) -> List[pathlib.WindowsPath]:
-    """Resolve multiple files."""
-    resolved_files = []
-    for file in files:
-        if windows_shortcuts.is_shortcut(file):
-            resolved_files.append(windows_shortcuts.read_shortcut(file))
-        else:
-            resolved_files.append(file.resolve())
-    return resolved_files
